removing extents.hpp and decoupling test_extents and test_strides
