+++ include/boost/numeric/ublas/tensor/dynamic_strides.hpp
@@ -132,68 +132,78 @@ public:
         std::swap(lhs._base   , rhs._base);
     }
 
-    const_reference operator[] (size_type p) const{
+    [[nodiscard]] inline
+    constexpr const_reference operator[] (size_type p) const{
         return _base[p];
     }
 
-    const_pointer data() const{
+    [[nodiscard]] inline
+    constexpr const_pointer data() const{
         return _base.data();
     }
 
-    const_reference at (size_type p) const{
+    [[nodiscard]] inline
+    constexpr const_reference at (size_type p) const{
         return _base.at(p);
     }
 
-    const_reference back () const{
+    [[nodiscard]] inline
+    constexpr const_reference back () const{
         return _base[_base.size() - 1];
     }
 
-    reference back (){
+    [[nodiscard]] inline
+    constexpr reference back (){
         return _base[_base.size() - 1];
     }
 
-    bool empty() const{
+    [[nodiscard]] inline
+    constexpr bool empty() const{
         return _base.empty();
     }
 
-    size_type size() const{
+    [[nodiscard]] inline
+    constexpr size_type size() const{
         return _base.size();
     }
 
-    template<class other_layout>
-    bool operator == (basic_strides<value_type, other_layout> const& b) const{
-        return b.base() == this->base();
-    }
-
-    template<class other_layout>
-    bool operator != (basic_strides<value_type, other_layout> const& b) const{
-        return b.base() != this->base();
-    }
-
-    bool operator == (basic_strides const& b) const{
-        return b._base == _base;
-    }
-
-    bool operator != (basic_strides const& b) const{
-        return b._base != _base;
-    }
-
-    const_iterator begin() const{
+    [[nodiscard]] inline
+    constexpr const_iterator begin() const{
         return _base.begin();
     }
 
-    const_iterator end() const{
+    [[nodiscard]] inline
+    constexpr const_iterator end() const{
         return _base.end();
     }
 
-    void clear() {
+    inline
+    constexpr void clear() {
         this->_base.clear();
     }
 
-    base_type const& base() const{
+    [[nodiscard]] inline
+    constexpr base_type const& base() const{
         return this->_base;
     }
 
+    template <class Strides, std::enable_if_t<is_strides_v<Strides>, int> = 0 >
+    [[nodiscard]] inline
+    constexpr bool operator==(Strides const& rhs) const noexcept{
+        static_assert(is_strides_v<Strides>,
+            "boost::numeric::ublas::operator==() : invalid type, type should be an extents");
+       if( this->size() != rhs.size() ){
+            return false;
+        }else{
+            return std::equal(this->begin(), this->end(), rhs.begin());
+        }
+    }
+
+    template <class Strides, std::enable_if_t<is_strides_v<Strides>, int> = 0 >
+    [[nodiscard]] inline
+    constexpr bool operator!=(Strides const& rhs) const noexcept{
+        return !( *this == rhs );
+    }
 
 protected:
     base_type _base;
@@ -206,14 +216,32 @@ protected:
 
 namespace boost::numeric::ublas{
     
-template <class L, class T> 
-struct is_strides<basic_strides<T,L>> : std::true_type {};
+    template <class L, class T> 
+    struct is_strides<basic_strides<T,L>> : std::true_type {};
 
-template <class T, class L>
-struct is_dynamic< basic_strides<T,L> > : std::true_type {};
+    template <class T, class L>
+    struct is_dynamic< basic_strides<T,L> > : std::true_type {};
 
-template <class T, class L>
-struct is_dynamic_rank< basic_strides<T, L> > : std::true_type {};
+    template <class T, class L>
+    struct is_dynamic_rank< basic_strides<T, L> > : std::true_type {};
+
+    namespace detail{
+        
+        /** @brief Partial Specialization of strides for basic_extents
+         *
+         *
+         * @tparam Layout either first_order or last_order
+         *
+         * @tparam T extents type
+         *
+         */
+        template <class Layout, class T>
+        struct strides_impl<basic_extents<T>, Layout>
+        {
+            using type = basic_strides<T, Layout>;
+        };
+        
+    } // detail
 
 } // namespace boost::numeric::ublas
