+++ include/boost/numeric/ublas/tensor/tensor.hpp
@@ -21,7 +21,9 @@
 #include <boost/numeric/ublas/tensor/algorithms.hpp>
 #include <boost/numeric/ublas/tensor/expression.hpp>
 #include <boost/numeric/ublas/tensor/expression_evaluation.hpp>
-#include <boost/numeric/ublas/tensor/extents.hpp>
+#include <boost/numeric/ublas/tensor/fixed_rank_extents.hpp>
+#include <boost/numeric/ublas/tensor/static_extents.hpp>
+#include <boost/numeric/ublas/tensor/dynamic_extents.hpp>
 #include <boost/numeric/ublas/tensor/strides.hpp>
 #include <boost/numeric/ublas/tensor/index.hpp>
 #include <boost/numeric/ublas/tensor/detail/type_traits.hpp>
