+++ test/tensor/Jamfile
@@ -57,6 +57,8 @@ test-suite boost-ublas-tensor-test
           unit_test_framework
     ]
     [ run test_fixed_rank_tensor.cpp
+          test_fixed_rank_extents.cpp
+          test_fixed_rank_strides.cpp
           test_fixed_rank_operators_arithmetic.cpp
           test_fixed_rank_operators_comparison.cpp
           test_fixed_rank_expression_evaluation.cpp
